Sprint_3 added fixes after Review 1
