+++ src/main/java/model/CourierLogin.java
@@ -1,10 +1,10 @@
-package com.example;
+package model;
 
 public class CourierLogin  {
     private String login;
     private String password;
 
-    CourierLogin(String login, String password){
+    public CourierLogin(String login, String password){
         this.login = login;
         this.password = password;
