+++ src/test/java/com/example/OrderAcceptTest.java
@@ -1,10 +1,14 @@
 package com.example;
 
+import api.CourierClient;
+import api.OrderClient;
 import io.qameta.allure.Description;
-import io.qameta.allure.Step;
 import io.qameta.allure.junit4.DisplayName;
 import io.restassured.RestAssured;
 import io.restassured.response.Response;
+import model.Courier;
+import model.CourierLogin;
+import model.Order;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
@@ -12,15 +16,13 @@ import org.junit.Test;
 import java.util.ArrayList;
 import java.util.List;
 
-import static org.hamcrest.CoreMatchers.equalTo;
-
-public class Test_OrderAccept {
-    private ArrayList<String> courierData ;
+public class OrderAcceptTest {
+    private ArrayList<String> courierData;
     private List<Integer> courierListToDelete = new ArrayList<>();
     private List<Integer> trackOrderListToDelete = new ArrayList<>();
     private Courier courier;
     private CourierLogin courierLogin;
-    private  Order order;
+    private Order order;
     private String login;
     private String password;
     private String firstName;
@@ -31,7 +33,7 @@ public class Test_OrderAccept {
     int signsQuantity;
 
     @Before
-    public void setUp(){
+    public void setUp() {
         RestAssured.baseURI = "http://qa-scooter.praktikum-services.ru/";
         fieldsSet = 1;
         signsQuantity = 10;
@@ -41,40 +43,27 @@ public class Test_OrderAccept {
         password = courierData.get(1);
         firstName = courierData.get(2);
 
-        courier = Test_CourierLogin.courierCreate(login, password, firstName);
-        courierId = Test_OrderListGet.getCourierIdCertainCourier(courier);
+        courier = CourierClient.courierCreate(login, password, firstName);
+        courierId = OrderClient.getCourierIdCertainCourier(courier);
         courierListToDelete.add(courierId);
 
-        trackOrder = Test_OrderListGet.getTrackOfOrderMetroStation("1");
+        trackOrder = OrderClient.getTrackOfOrderMetroStation("1");
         trackOrderListToDelete.add(trackOrder);
-        id  = Test_OrderListGet.getOrderId(trackOrder);
+        id = OrderClient.getOrderId(trackOrder);
     }
 
     @Test
     @DisplayName("COurier accept an Order. Result 200")
     @Description("Create a courier, Create an Order, Accept An Order")
     public void testCourierAcceptAnOrder() {
-        Response response = Test_OrderListGet.courierAcceptOrder(id,courierId);
+        Response response = OrderClient.courierAcceptOrder(id, courierId);
         response.then().assertThat().statusCode(200);
     }
 
     @After
     public void standClear() {
-        Test_CourierCreate.deleteAllCouriers(courierListToDelete);
-        deleteAllOrders(trackOrderListToDelete);
+        CourierClient.deleteAllCouriers(courierListToDelete);
+        OrderClient.deleteAllOrders(trackOrderListToDelete);
     }
-
-    @Step("Order cancel doesn't work properly.")
-    public static void deleteAllOrders(List<Integer> trackOrdersListToDelete) {
-
-        for(Integer orderTrack: trackOrdersListToDelete) {
-            Response responseCourierDelete = Test_OrderCreateColorParametrized.orderCancel(orderTrack);
-
-        }
-
-        trackOrdersListToDelete.clear();
-    }
-
-
 
 }
